+++ src/components/Calendar/Calendar.tsx
@@ -1,4 +1,4 @@
-// Tremor Raw Calendar [v0.0.2]
+// Tremor Raw Calendar [v0.0.3]
 
 "use client"