+++ src/components/Button/changelog.md
@@ -1,5 +1,11 @@
 # Tremor Raw Button Changelog
 
+## 0.1.1
+
+### Changes
+
+- Fix: Default height
+
 ## 0.1.0
 
 ### Changes