+++ src/components/Popover/changelog.md
@@ -1,5 +1,11 @@
 # Tremor Raw Popover Changelog
 
+## 0.0.2
+
+### Changes
+
+- Fix: Border color
+
 ## 0.0.1
 
 ### Changes