+++ src/components/Select/changelog.md
@@ -1,5 +1,12 @@
 # Tremor Raw Select Changelog
 
+## 0.0.2
+
+### Changes
+
+- Fix: Border color
+- Fix: Placeholder color
+
 ## 0.0.1
 
 ### Changes